adds 'busy' constraint
As I continue on this project, i'll realize more constraints and I'll update
this list accordingly. It feels fair to make it discoverable about how the
'constraints' list grows over time. Things that might be obvious at the end
of this project might be difficult to appreciate or even believe at the
beginning, but I want to mark a path for how they may have gotten on thise
list.