making final extension mostly pass - sorts on two attributes
I've no doubt I could figure out how to make it sort :asc and :desc
but I'm not really interested enough in this to continue. I've soaked
this project for it's learning outcomes, and outlined every step of
the way in excruciating detail.

You, dear Turing student, are PROBABLY able to recreate this entire
class, almost exactly as I've written it, if you give it some time.

If you do that, you'll have learned a lot.